--- command/delete-user_test.go
@@ -1,11 +1,0 @@
-package command
-
-import (
-	"testing"
-
-	"github.com/mitchellh/cli"
-)
-
-func TestDeleteUserCommand_implement(t *testing.T) {
-	var _ cli.Command = &DeleteUserCommand{}
-}
